Move ResolveImports into module, update modules and views.
